move nest_fields back to utils
