+++ src/agoradatatools/etl/transform/custom/distribution_data.py
@@ -1,4 +1,5 @@
 import pandas as pd
+import numpy as np
 
 
 def calculate_distribution(df: pd.DataFrame, col: str, is_scored, upper_bound) -> dict:
